requirements.txt
jsonpath-ng>=1.5.3
requests>=2.28.0
websocket-client>=1.3.2
python-dotenv>=0.20.0
